requirements.txt
asgiref==3.8.1
Django==5.1.7
django-unfold==0.53.0
sqlparse==0.5.3
